Bug fix: `View.writeRune` appends stray `\x00` when writing to the end of a line
The cause was that the line array was extended twice when writing at the end of
a line - once in the first `if` clause because the cursor is beyond the end of
the line, and then again in the second `if` clause dealing with the `Overwrite`
condition.

This change makes the two `if` clauses mutually exclusive.