Added context.Context support
Added the method `MainLoopWithContext`, which allows for callers to supply their own context for running a `Gui`. I made this addition because of a project I'm personally working on (and hope to add to the README list soon). My project is heavily concurrent and their exists an abstraction layer between the `Gui` and a backend. The abstraction layer handles many things, mainly monitoring `os.Signal`s and others, which means I couldn't/(don't want to) allow `Gui` to handle termination. Currently, there is no way to solve my problem. Hence, this pull request adds the solution to my problem and hopefully brings even more possibilities to what can be done with `gocui`.