Add frame customization
FrameColor can be specified for View.
TitleColor can be specified too and is applied to title and subtitle.
This color will be used if no Highlight or no current View is displayed.

FrameRunes can be specified for View allowing to change frame style.

Added custom_frame.go example for demonstration