Change colors to mimic tcell/v2 colors
Colors are changed to be represented exactly as tcell/v2 colors.
User can use tcell functions and just wrap them in Attribute() to create
colors.

Backward compatibility for application using termbox-go coloring system.
If colors are in 256 range and are not marked as valid colors
(which would be default for previously build applications), they will
be translated to tcell, by subtracting 1 from it.