If selected line is outside, move it to the middle of the view
... instead of scrolling just as much as necessary to make it visible again. For
example, when the cursor is on the last line and I press down-arrow, it would
previously only scroll by one line; now I scrolls by half a page. I prefer this
because it lets me see more of what's coming in the direction that I'm moving
to.

It doesn't only affect arrow keys; it also applies to selecting the next search
hit using "n" in the commits panel. I like it here too, it makes it easier to
see how many more hits are coming below.