Add FocusHandler
I decided to make this a global handler that can be set only once. Alternatively
we could consider dispatching focus events to the focused view similar to how we
do with key bindings; but then we'd need a way to dispatch it to parent views
that might also be interested. Since we don't need this functionality (yet?) I
went with the simplest solution for now.