Only clear runes when view is visible
Calling v.draw() on a view that is invisible would clear the views area. This
wasn't a problem so far because currently the only caller of draw() is
gui.draw(), which checks for v.Visible itself and doesn't call draw() for
invisible views. We want to introduce a new caller of draw in the next commit
though, and we don't want to duplicate the Visible check there, so fix it to
behave correctly.